don't create gravity tasks when there are no gravity particles.
Former-commit-id: 67aa9f90856740bbfec0f4a0768aefead4ce51c5

+++ src/part.h
@@ -68,7 +68,7 @@ struct gpart {
     union {
     
         /* Particle ID. */
-        unsigned long long id;
+        size_t id;
 
         /* Pointer to corresponding SPH part. */
         struct part *part;
